pyproject.toml
[tool.poetry]
name = "web-tracker-backend"
version = "0.1.0"
description = "GESIS Backend for Web Tracking Tool"
authors = ["geomario <marm1984@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "web_tracker_backend" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-stubs = "^4.2.6"
psycopg2 = "^2.9.9"
types-psycopg2 = "^2.9.21.15"
drf-spectacular = "^0.27.1"
uwsgi = "^2.0.25"
django-simple-history = "^3.5.0"
celery = "^5.3.6"
redis = "^5.0.3"
flower = "^2.0.1"
python-logstash = "^0.4.8"
cryptography = "^44.0.1"
pygraphviz = "^1.11"
elasticsearch = "^8.17.0"
elasticsearch-dsl = "^8.17.0"
django-cors-headers = "^4.9.0"
celery-stubs = "^0.1.3"
django-extensions = "^4.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-bugbear = "^23.9.16"
flake8-isort = "^6.1.0"
flake8-annotations = "^3.0.1"
black = "^23.10.1"
pytest = "^7.4.3"
mypy = "^1.6.1"
commitizen = "^3.12.0"
pre-commit = "^3.5.0"
djangorestframework-stubs = "^3.14.4"
django-stubs = "^4.2.6"
pylint-django = "^2.5.5"
pylint = "^3.3.3"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
mypy_path = "app"

[tool.django-stubs]
django_settings_module = "app.settings"

[tool.isort]
profile = "black"
remove_imports = "from __future__ import annotations"
force_single_line = true

[tool.black]
target-version = ["py310"]
